+++ src/minecraft/ip/industrialProcessing/multiblock/core/extend/TileEntityMultiblockCoreInvWorkerPowered.java
@@ -19,10 +19,10 @@ import net.minecraftforge.common.ForgeDirection;
 public abstract class TileEntityMultiblockCoreInvWorkerPowered extends TileEntityMultiblockCoreInvWorker implements IPoweredMachine {
 
     public TileEntityMultiblockCoreInvWorkerPowered(StructureMultiblock structure, TierCollection tierRequirments, RecipesMultiblock recipes, LocalDirection powerInput, int powerCapacity, int maxWorkSpeed) {
-	super(structure, tierRequirments, recipes);
-	this.powerInputSide = powerInput;
-	this.maxWorkSpeed = maxWorkSpeed;
-	this.storage = new SimplePowerStorage(powerCapacity);
+        super(structure, tierRequirments, recipes);
+        this.powerInputSide = powerInput;
+        this.maxWorkSpeed = maxWorkSpeed;
+        this.storage = new SimplePowerStorage(powerCapacity);
     }
 
     private LocalDirection powerInputSide;
@@ -31,66 +31,66 @@ public abstract class TileEntityMultiblockCoreInvWorkerPowered extends TileEntit
 
     @Override
     public void doWork() {
-	Recipe recipe = ((RecipeMultiblockWorker) getWorker()).getCurrentRecipe();
+        Recipe recipe = ((RecipeMultiblockWorker) getWorker()).getCurrentRecipe();
 
-	if (recipe != null) {
-	    int amount = PowerWorkerHelper.getWork(this.storage, this.maxWorkSpeed);
-	    int maxWork = this.storage.drainPower(amount, false);
+        if (recipe != null) {
+            int amount = PowerWorkerHelper.getWork(this.storage, this.maxWorkSpeed);
+            int maxWork = (int) this.storage.drainPower(amount, false);
 
-	    if (recipe instanceof IPowerRecipe) {
-		IPowerRecipe powerRecipe = (IPowerRecipe) recipe;
-		if (powerRecipe.getPowerRequired() > 0)
-		    maxWork /= powerRecipe.getPowerRequired();
-		else
-		    maxWork = this.maxWorkSpeed;
-	    } else
-		maxWork = this.maxWorkSpeed;
+            if (recipe instanceof IPowerRecipe) {
+                IPowerRecipe powerRecipe = (IPowerRecipe) recipe;
+                if (powerRecipe.getPowerRequired() > 0)
+                    maxWork /= powerRecipe.getPowerRequired();
+                else
+                    maxWork = this.maxWorkSpeed;
+            } else
+                maxWork = this.maxWorkSpeed;
 
-	    int maxPower = work(maxWork);
-	    if (recipe instanceof IPowerRecipe) {
-		IPowerRecipe powerRecipe = (IPowerRecipe) recipe;
-		maxPower *= powerRecipe.getPowerRequired();
-		this.storage.drainPower(maxPower, true);
-	    }
-	}
+            int maxPower = work(maxWork);
+            if (recipe instanceof IPowerRecipe) {
+                IPowerRecipe powerRecipe = (IPowerRecipe) recipe;
+                maxPower *= powerRecipe.getPowerRequired();
+                this.storage.drainPower(maxPower, true);
+            }
+        }
     }
 
     @Override
     public void readFromNBT(NBTTagCompound nbt) {
-	super.readFromNBT(nbt);
-	this.storage.readFromNBT(nbt);
+        super.readFromNBT(nbt);
+        this.storage.readFromNBT(nbt);
     }
 
     @Override
     public void writeToNBT(NBTTagCompound nbt) {
-	super.writeToNBT(nbt);
-	this.storage.writeToNBT(nbt);
+        super.writeToNBT(nbt);
+        this.storage.writeToNBT(nbt);
     }
 
     @Override
     public float getResistance(ForgeDirection side, float voltage) {
-	if (canAcceptPower(side)) {
-	    return PowerHelper.getResistanceForStorage(this.storage);
-	}
-	return Float.POSITIVE_INFINITY;
+        if (canAcceptPower(side)) {
+            return PowerHelper.getResistanceForStorage(this.storage);
+        }
+        return Float.POSITIVE_INFINITY;
     }
 
     @Override
     public void applyPower(ForgeDirection side, float coulombs, float voltage) {
-	if (canAcceptPower(side)) {
-	    int joules = (int) PowerHelper.getEnergy(coulombs, voltage);
-	    this.storage.fillPower(joules, true);
-	}
+        if (canAcceptPower(side)) {
+            int joules = (int) PowerHelper.getEnergy(coulombs, voltage);
+            this.storage.fillPower(joules, true);
+        }
     }
 
     @Override
     public boolean canAcceptPower(ForgeDirection side) {
-	return this.powerInputSide == DirectionUtils.getLocalDirection(side, getForwardDirection());
+        return this.powerInputSide == DirectionUtils.getLocalDirection(side, getForwardDirection());
     }
 
     @Override
     public IPowerStorage getMainPowerStorage() {
-	return this.storage;
+        return this.storage;
     }
 
 }
